+++ classes/Policy.php
@@ -26,8 +26,9 @@ class Policy
         $properties = $this->properties($class);
         $table      = $this->table($class);
         $key        = $this->key($class);
+        $factory    = $this->factory($class);
 
-        $repository = new Repository($connection, $class, $table, $key, $properties);
+        $repository = new Repository($connection, $class, $table, $key, $properties, $factory);
         if ($softDelete = $this->softDelete($class)) {
             $repository = new SoftDelete($repository, $softDelete, $properties[$softDelete]);
         }
@@ -152,6 +153,17 @@ class Policy
     public function skip(Property $property): bool
     {
         return $property->static() || preg_match('/\\s@nostorage\\s/', $property->comment());
+    }
+
+    /**
+     * Get factory method
+     *
+     * @param string $class
+     * @return callable|null
+     */
+    public function factory(string $class)
+    {
+        return method_exists($class, 'createFromArray') ? [$class, 'createFromArray'] : null;
     }
 
     /**