+++ tests/Helper/Type.php
@@ -9,4 +9,17 @@ class Type
 
     /** @var string */
     public $name;
+
+    /**
+     * @param array $array
+     * @return Type
+     */
+    public static function createFromArray(array $array): Type
+    {
+        $type = new Type();
+        $type->id   = (int) $array['id'];
+        $type->name = (string) $array['name'];
+
+        return $type;
+    }
 }